requirements.txt
google-adk>=1.1.1
pytest>=7.3.1
python-dotenv
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
asyncio>=3.4.3
google-generativeai>=0.6.0
mcp[cli]>=1.9.0
gitignore-parser>=0.1.0
